sts_rasa_twilio/requirements.txt
rasa==3.6.21
twilio
python-dotenv
requests
deepgram-sdk
websockets
fastapi
uvicorn
python-socketio